test_perception_matching/pyproject.toml
[project]
name = "test-perception-matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.0",
    "opencv-python==4.8.1.78",
    "opencv-python-headless==4.8.1.78",
    "pillow>=11.2.1",
    "psutil==7.0.0",
    "psycopg==3.2.6",
    "psycopg-binary==3.2.6",
    "pydantic>=2.11.3",
    "qdrant-client>=1.12.1",
    "quadrant>=1.0",
    "scikit-learn==1.3.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
